Safari-609.1.11.1/LayoutTests/storage/websql/openDatabase-deprecation.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <title>openDatabase deprecation</title>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
assert_equals(Object.keys(window).indexOf("openDatabase"), -1);
window.openDatabase(null, null, null, null);
done();
        </script>
    </body>
</html>
